% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_capture_hist.R
\name{calc_detection}
\alias{calc_detection}
\title{Calculate Trap Detection Probabilities}
\usage{
calc_detection(ntraps, homeranges, traps, tau, p)
}
\arguments{
\item{ntraps}{number of traps}

\item{homeranges}{homeranges for all individuals}

\item{traps}{locations of all traps}

\item{tau}{movement parameter for species}

\item{p}{capture probability for species}
}
\value{
trap_detec 2D array of detection probabilities for each individual i
and each trap j.
}
\description{
Function that calculates the detection probabilities for each
individual i and each trap j
}
\examples{

# plot individuals and their camera trap detection probabilities
# labelled for trap j. Blue label is individual i number.

# number of individuals in population
n <- 5
homeranges <- create_homeranges(n)

# number of camera traps
ntraps <- 3

# movement parameter
tau <- 0.55

# species capture probability
p <- 0.6

traps <- define_traps(ntraps)
trap_detec <- calc_detection(ntraps, homeranges, traps, tau, p)

for (j in 1:ntraps) {
  par(pty = "s", mar = c(5, 4, 4, 8), xpd = TRUE)
  plot(homeranges, xlim = c(0, 1), ylim = c(0, 1))
  points(x = traps[j, 1], y = traps[j, 2], col = "red", pch = 15)
  text(
    x = traps[j, 1], y = traps[j, 2], labels = paste(c("Trap", j), collapse = " "),
    cex = 1.5, font = 2, pos = 3
  )
  text(homeranges, labels = round(trap_detec[, j], 2), cex = 1, font = 2, pos = 3)
  text(homeranges, labels = rownames(homeranges), cex = 1, font = 2, pos = 2, col = "blue")
  legend("topright",
    inset = c(-.5, 0), c("homerange", "camera trap"), cex = .8,
    col = c("black", "red"), pch = c(1, 15)
  )
}
}
