% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presence_absence.R
\name{calc_prop}
\alias{calc_prop}
\title{Calculate Proportion of Time Spent in Grid Cell}
\usage{
calc_prop(x, y, x0, x1, y0, y1, tau)
}
\arguments{
\item{x}{The x coordinate of the homerange for an individual}

\item{y}{The y coordinate of the homerange for an individual}

\item{x0}{The starting x value for the grid cell}

\item{x1}{The ending x value for the grid cell}

\item{y0}{The starting y value for the grid cell}

\item{y1}{The ending y value for the grid cell}

\item{tau}{The movement parameter for the species, lower is less movement}
}
\value{
A value between 0 and 1 indicating the proportion of time that the
individual spends in the given grid cell
}
\description{
Calculate Proportion of Time Spent in Grid Cell
}
\examples{
# show high proportion when homerange in center of grid cell
calc_prop(0.25 / 2, 0.25 / 2, 0, 0.25, 0, 0.25, 0.1)
}
