% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_capture_hist.R
\name{create_cap_hist}
\alias{create_cap_hist}
\title{Create Capture History}
\usage{
create_cap_hist(homeranges, nsample_cap, camera_trap_full)
}
\arguments{
\item{homeranges}{homeranges for each individual}

\item{nsample_cap}{number of capture-recapture sampling occasions}

\item{camera_trap_full}{a 3D array of whether each individual i was seen by
trap j at period k}
}
\value{
capture_hist a 2D array of capture histories for each individual
}
\description{
Function that collapses the 3D array camera_trap_full (individual i, trap j, period k)
on the jth dimension to obtain the capture history for each individual i.
}
