% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_homeranges.R
\name{create_homeranges}
\alias{create_homeranges}
\title{Create Homeranges}
\usage{
create_homeranges(n)
}
\arguments{
\item{n}{number of individuals}
}
\value{
a 2D array of the homeranges for each individual
}
\description{
Function that assigns homeranges to n individuals using a Poisson
random process
}
\details{
Assumes a study area that is a unit square (0,1) x (0,1)
}
\examples{
n <- 10
homeranges <- create_homeranges(n)
# plotting homeranges for validation
par(pty = "s")
plot(homeranges, xlim = c(0, 1), ylim = c(0, 1))
text(homeranges, labels = rownames(homeranges), cex = 1.5, font = 2, pos = 3)
}
