% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_capture_hist.R
\name{define_traps}
\alias{define_traps}
\title{Define Trap Locations}
\usage{
define_traps(ntraps)
}
\arguments{
\item{ntraps}{the number of camera traps to place}
}
\value{
traps a 2D array of trap locations in x,y coordinates
}
\description{
Function that places camera traps in the study area.
}
\details{
Uses a Poisson process where x, y coordinates of each trap
are independent uniform random variables
}
\examples{
n <- 5
homeranges <- create_homeranges(n)
ntraps <- 3
traps <- define_traps(ntraps)
# plot traps on study area with homeranges included
par(pty = "s", mar = c(5, 4, 4, 8), xpd = TRUE)

plot(homeranges, xlim = c(0, 1), ylim = c(0, 1))
points(traps, col = "red", pch = 15)
legend("topright",
  inset = c(-.5, 0), c("homerange", "camera trap"),
  cex = .8, col = c("black", "red"), pch = c(1, 15)
)
}
