% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distance}
\alias{distance}
\title{Distance function}
\usage{
distance(a, b)
}
\arguments{
\item{a}{point a in vector form}

\item{b}{point b in vector form}
}
\value{
euclidean distance between point a and point b
}
\description{
Distance function
}
\examples{
a <- c(2, 6)
b <- c(3, 5)
distance(a, b)
}
