% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presence_absence.R
\name{generate_presence_absence}
\alias{generate_presence_absence}
\title{Generate Presence-Absence Data}
\usage{
generate_presence_absence(nsites, nsample_pa, tau, sigma, delta, homeranges)
}
\arguments{
\item{nsites}{number of sites in study area}

\item{nsample_pa}{number of presence-absence sampling occasions}

\item{tau}{movement parameter of species}

\item{sigma}{the marker deposition rate}

\item{delta}{the time between sampling occasions}

\item{homeranges}{array of homerange centers for all individuals}
}
\value{
presence_absence a 2D array (nsites rows by nsample_pa columns)
of 0s and 1s where 1 indicates a detection for a site on a period.
}
\description{
Function that generates the presence-absence array.
}
\examples{

# number of sites
nsites <- 4**2

# number of individuals in study area
n <- 10

# amount of movement for the species
tau <- 0.25

# marker deposition rate
sigma <- 1

# time between sampling occasions (presence-absence)
delta <- 2

# number of sampling occasions in presence-absence
nsample_pa <- 7

# homeranges for the n individuals
homeranges <- create_homeranges(n)

presence_absence <- generate_presence_absence(
  nsites, nsample_pa, tau, sigma,
  delta, homeranges
)
}
