% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{half_norm_detec}
\alias{half_norm_detec}
\title{Half-Normal Detection Function}
\usage{
half_norm_detec(d, tau)
}
\arguments{
\item{d}{distance of homerange center to trap}

\item{tau}{movement parameter for species}
}
\value{
probability of detection for individual by camera trap
}
\description{
A half-normal detection function models the capture probability of individual i by trap j:\if{html}{\out{
<p>p<sub>ij</sub> = p e<sup>(-d<sup>2</sup><sub>ij</sub>/&tau;<sup>2</sup>)</sup></p>
<p>where <em>p</em> is the capture probability for a camera trap at distance 0 from the home range
centre, <em>&tau;</em> is the movement parameter for the species, and <em>d<sub>ij</sub></em>
is the distance between the home-range centre <em>i</em> and the camera location <em>j</em>.</p>
}}
}
\examples{
# example on species with tau=0.1
trap <- c(0.25, 0.5)
homerange <- c(0.26, 0.6)
tau <- 0.1

d <- distance(trap, homerange)
detect_prob <- half_norm_detec(d, tau)
}
