% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_data}
\alias{output_data}
\title{Output Data}
\usage{
output_data(presence_absence, capture_hist, id)
}
\arguments{
\item{presence_absence}{returned data from generate_presence_absence}

\item{capture_hist}{returned data from generate_capture_hist}

\item{id}{a unique identifier for the outputted files}
}
\value{
Outputs capture history text and presence absence csv files
}
\description{
Function that outputs the capture-recapture and presence-absence data
into text and csv formats respectively.
}
