% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presence_absence.R
\name{setup_grid}
\alias{setup_grid}
\title{Setup Grid Area
function that divides the study area (a unit square) into a grid of equally
divided regions.}
\usage{
setup_grid(nsites)
}
\arguments{
\item{nsites}{number of grid cells required}
}
\value{
regions, a 2-d array with grid cells defined as x0,x1,y0,y1
bounding boxes
}
\description{
Setup Grid Area
function that divides the study area (a unit square) into a grid of equally
divided regions.
}
\examples{

# number of individuals
n <- 5
homeranges <- create_homeranges(n)
nsites <- 16
regions <- setup_grid(nsites)
#
# plot grid cells on study area with homerange centers and individuals labeled
par(pty = "s")
plot(homeranges, xlim = c(0, 1), ylim = c(0, 1))
m <- sqrt(nsites)
for (i in 0:m) {
  abline(v = i / m)
  abline(h = i / m)
}
# individuals labeled from 1 to n
text(homeranges, labels = rownames(homeranges), cex = 1.5, font = 2, pos = 3)
}
