% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_presence_absence.R
\name{sim_presence_absence}
\alias{sim_presence_absence}
\title{Simulate Presence-Absence Data}
\usage{
sim_presence_absence(
  nsites,
  nsample_pa,
  homeranges,
  regions,
  tau,
  sigma,
  delta
)
}
\arguments{
\item{nsites}{number of regions in the study area}

\item{nsample_pa}{number of sampling occasions}

\item{homeranges}{the homeranges of the individuals as 2D array of coordinates}

\item{regions}{the subareas of the study area, as 2D array of (x0,x1,y0,y1)
bounding boxes}

\item{tau}{the movement parameter for the species, higher is more movement}

\item{sigma}{the marker deposition rate (integer)}

\item{delta}{the time between sampling occasions (integer)}
}
\value{
presence_absence a 2D array (nsites rows by nsample_pa columns)
of 0s and 1s where 1 indicates a detection for a site and a period.
}
\description{
Function that simulates presence-absence data for each region and for
each period.
}
\details{
Assumes that the amount of time spent by each individual
in each region is proportional to the probability of detection.
}
