% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_capture_hist.R
\name{sim_trap_interac}
\alias{sim_trap_interac}
\title{Simulate Trap Interactions}
\usage{
sim_trap_interac(homeranges, ntraps, nsample_cap, trap_detec)
}
\arguments{
\item{homeranges}{homeranges of the individuals}

\item{ntraps}{number of camera traps}

\item{nsample_cap}{number of capture-recapture sampling occasions}

\item{trap_detec}{2D array of detection probabilities for each individual i
and each trap j.}
}
\value{
camera_trap_full a 3D array of whether each individual i was seen by
trap j at period k
}
\description{
Function that simulates interactions of individuals with the camera traps.
}
\examples{

# movement parameter
tau <- 0.55

# number of individuals in population
n <- 5

# species capture probability
p <- 0.6

homeranges <- create_homeranges(n)
ntraps <- 3

traps <- define_traps(ntraps)
trap_detec <- calc_detection(ntraps, homeranges, traps, tau, p)

# number of capture-recapture sampling occasions
nsample_cap <- 5

camera_trap_full <- sim_trap_interac(homeranges, ntraps, nsample_cap, trap_detec)

# plot number of times each individual was seen by each trap
trap_counts <- rowSums(camera_trap_full, dims = 2)

for (j in 1:ntraps) {
  barplot(trap_counts[, j],
    names.arg = rownames(homeranges),
    main = paste(c("Trap", j), collapse = " "), xlab = "individual", ylab = "count"
  )
}
}
