% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtrapoccR-package.R
\docType{package}
\name{simtrapoccR-package}
\alias{simtrapoccR}
\alias{simtrapoccR-package}
\title{simtrapoccR: Simulate Camera Trap Capture-Recapture and Presence-Absence Occupancy Data}
\description{
The 'simtrapoccR' package allows you to generate capture-recapture history matrices and presence-absence history matrices for feeding into population estimation statistical models. Information on how the data are generated can be found at https://github.com/jackthomas818/masters-project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jackthomas818/simtrapoccR}
  \item Report bugs at \url{https://github.com/jackthomas818/simtrapoccR/issues}
}

}
\author{
\strong{Maintainer}: Jack Thomas \email{jack.h.thomas567@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
